# Tweak: Optimize global.css not to generate styles for unused widgets [ED-15523]