--- a/core/files/css/global-css.php
+++ b/core/files/css/global-css.php
@@ -1,5 +1,5 @@
 <?php
-// TODO: Delete this file in v3.26.0
+// TODO: Delete this file in v3.26.0 - It is not in use anymore.
 namespace Elementor\Core\Files\CSS;
 
 use Elementor\Core\Kits\Manager;