--- a/core/files/css/post.php
+++ b/core/files/css/post.php
@@ -312,7 +312,13 @@ class Post extends Base {
 	private function render_element_styles( Element_Base $element ) {
 		$element_settings = $element->get_settings();
 
-		$this->add_controls_stack_style_rules( $element, $this->get_style_controls( $element, null, $element->get_parsed_dynamic_settings() ), $element_settings, [ '{{ID}}', '{{WRAPPER}}' ], [ $element->get_id(), $this->get_element_unique_selector( $element ) ] );
+		$this->add_controls_stack_style_rules(
+			$element,
+			$this->get_style_controls( $element, null, $element->get_parsed_dynamic_settings() ),
+			$element_settings,
+			[ '{{ID}}', '{{WRAPPER}}' ],
+			[ $element->get_id(), $this->get_element_unique_selector( $element ) ]
+		);
 	}
 
 	private function render_element_global_styles( Element_Base $element ) {