--- a/core/files/manager.php
+++ b/core/files/manager.php
@@ -4,7 +4,6 @@ namespace Elementor\Core\Files;
 use Elementor\Core\Base\Document as Document_Base;
 use Elementor\Core\Base\Elements_Iteration_Actions\Assets;
 use Elementor\Core\Common\Modules\Ajax\Module as Ajax;
-use Elementor\Core\Files\CSS\Global_CSS;
 use Elementor\Core\Files\CSS\Post as Post_CSS;
 use Elementor\Core\Page_Assets\Data_Managers\Base as Page_Assets_Data_Manager;
 use Elementor\Core\Responsive\Files\Frontend;
@@ -117,7 +116,6 @@ class Manager {
 		delete_post_meta_by_key( Document_Base::CACHE_META_KEY );
 		delete_post_meta_by_key( Assets::ASSETS_META_KEY );
 
-		delete_option( Global_CSS::META_KEY );
 		delete_option( Frontend::META_KEY );
 
 		$this->reset_assets_data();