--- a/includes/frontend.php
+++ b/includes/frontend.php
@@ -5,7 +5,6 @@ use Elementor\Core\Base\App;
 use Elementor\Core\Base\Elements_Iteration_Actions\Assets;
 use Elementor\Core\Frontend\Render_Mode_Manager;
 use Elementor\Core\Responsive\Files\Frontend as FrontendFile;
-use Elementor\Core\Files\CSS\Global_CSS;
 use Elementor\Core\Files\CSS\Post as Post_CSS;
 use Elementor\Core\Files\CSS\Post_Preview;
 use Elementor\Core\Responsive\Responsive;
@@ -692,8 +691,6 @@ class Frontend extends App {
 			do_action( 'elementor/frontend/after_enqueue_styles' );
 
 			if ( ! Plugin::$instance->preview->is_preview_mode() ) {
-				$this->parse_global_css_code();
-
 				$post_id = get_the_ID();
 				// Check $post_id for virtual pages. check is singular because the $post_id is set to the first post on archive pages.
 				if ( $post_id && is_singular() ) {
@@ -1062,20 +1059,6 @@ class Frontend extends App {
 
 		$this->fonts_to_enqueue[] = $font;
 		$this->registered_fonts[] = $font;
-	}
-
-	/**
-	 * Parse global CSS.
-	 *
-	 * Enqueue the global CSS file.
-	 *
-	 * @since 1.2.0
-	 * @access protected
-	 */
-	protected function parse_global_css_code() {
-		$scheme_css_file = Global_CSS::create( 'global.css' );
-
-		$scheme_css_file->enqueue();
 	}
 
 	/**