… tests/elements-regression/tests/elements-regression.test.ts-snapshots/text-editor-mobile-published-linux.png
… tests/elements-regression/tests/elements-regression.test.ts-snapshots/text-editor-published-linux.png
… tests/elements-regression/tests/elements-regression.test.ts-snapshots/text-editor-tablet-published-linux.png