**Page ID 2220 content:**
Container
Heading
Button
Nested Tabs
Progress Bar

**post-2220.css BEFORE (with a separate big global.css file loaded):**

`.elementor-2220 .elementor-element.elementor-element-9fc553c{--display:flex;--background-transition:0.3s;}.elementor-2220 .elementor-element.elementor-element-d1c0690{--display:flex;--background-transition:0.3s;}.elementor-2220 .elementor-element.elementor-element-4b01413 .elementor-button{background-color:var( --e-global-color-secondary );border-radius:2px 2px 2px 2px;}.elementor-2220 .elementor-element.elementor-element-ea19680{--display:flex;--background-transition:0.3s;}.elementor-2220 .elementor-element.elementor-element-eb04b9e{--display:flex;--background-transition:0.3s;}.elementor-2220 .elementor-element.elementor-element-f3fdbbe{--display:flex;--background-transition:0.3s;}.elementor-2220 .elementor-element.elementor-element-f663e4a{--n-tabs-heading-wrap:wrap;--n-tabs-heading-overflow-x:initial;--n-tabs-title-white-space:initial;}.elementor-2220 .elementor-element.elementor-element-11258a8{--display:flex;--background-transition:0.3s;}:root{--page-title-display:none;}`

**post-2220.css AFTER (contains the global css as well - for the widgets on the page):**

`.elementor-2220 .elementor-element.elementor-element-9fc553c{--display:flex;--background-transition:0.3s;}.elementor-widget-heading .elementor-heading-title{color:var( --e-global-color-primary );font-family:var( --e-global-typography-primary-font-family ), Sans-serif;font-weight:var( --e-global-typography-primary-font-weight );}.elementor-2220 .elementor-element.elementor-element-d1c0690{--display:flex;--background-transition:0.3s;}.elementor-widget-button .elementor-button{font-family:var( --e-global-typography-accent-font-family ), Sans-serif;font-weight:var( --e-global-typography-accent-font-weight );background-color:var( --e-global-color-accent );}.elementor-2220 .elementor-element.elementor-element-4b01413 .elementor-button{background-color:var( --e-global-color-secondary );border-radius:2px 2px 2px 2px;}.elementor-widget-nested-tabs.elementor-widget-n-tabs > .elementor-widget-container > .e-n-tabs[data-touch-mode="false"] > .e-n-tabs-heading > .e-n-tab-title[aria-selected="false"]:hover{background:var( --e-global-color-accent );}.elementor-widget-nested-tabs.elementor-widget-n-tabs > .elementor-widget-container > .e-n-tabs > .e-n-tabs-heading > .e-n-tab-title[aria-selected="true"], .elementor-widget-nested-tabs.elementor-widget-n-tabs > .elementor-widget-container > .e-n-tabs[data-touch-mode="true"] > .e-n-tabs-heading > .e-n-tab-title[aria-selected="false"]:hover{background:var( --e-global-color-accent );}.elementor-widget-nested-tabs.elementor-widget-n-tabs > .elementor-widget-container > .e-n-tabs > .e-n-tabs-heading > :is( .e-n-tab-title > .e-n-tab-title-text, .e-n-tab-title ){font-family:var( --e-global-typography-accent-font-family ), Sans-serif;font-weight:var( --e-global-typography-accent-font-weight );}.elementor-2220 .elementor-element.elementor-element-ea19680{--display:flex;--background-transition:0.3s;}.elementor-2220 .elementor-element.elementor-element-eb04b9e{--display:flex;--background-transition:0.3s;}.elementor-2220 .elementor-element.elementor-element-f3fdbbe{--display:flex;--background-transition:0.3s;}.elementor-2220 .elementor-element.elementor-element-f663e4a{--n-tabs-heading-wrap:wrap;--n-tabs-heading-overflow-x:initial;--n-tabs-title-white-space:initial;}.elementor-2220 .elementor-element.elementor-element-11258a8{--display:flex;--background-transition:0.3s;}.elementor-widget-progress .elementor-progress-wrapper .elementor-progress-bar{background-color:var( --e-global-color-primary );}.elementor-widget-progress .elementor-title{color:var( --e-global-color-primary );font-family:var( --e-global-typography-text-font-family ), Sans-serif;font-weight:var( --e-global-typography-text-font-weight );}:root{--page-title-display:none;}`

This was tested with 3rd Party widgets as well.